requirements.txt
Django==4.2
pillow
psycopg2-binary
ipython
black
redis
python-dotenv
flake8
isort
djangorestframework
django-filter
djangorestframework-simplejwt
coverage
drf-yasg
django-cors-headers
requests
stripe
celery
django-celery-beat
drf-spectacular
eventlet
gunicorn